# Improve performance on distinct