@Ekman 2020-06-17
I had some colleagues point out that this would be a breaking change since `in_array` can take an object, while `isset` can't.

An alternative solution would be to do something like this:

```php
function distinct($collection)
{
    $generatorFactory = function () use ($collection) {
        $distinctValues = [];
        $distinctObjects = new \SplObjectStorage();

        foreach ($collection as $key => $value) {
            if (is_object($value)) {
                if (!$distinctObjects->contains($value)) {
                    $distinctObjects->attach($value);
                    yield $key => $value;
                }
            } else {
                if (!isset($distinctValues[$value])) {
                    $distinctValues[$value] = true;
                    yield $key => $value;
                }
            }
        }
    };

    return new Collection($generatorFactory);
}
```
Let me know what you think.