@jdreesen 2020-06-17
The problem is, that `in_array` cannot only handle `objects`, but even `array`s (see: https://www.php.net/manual/en/function.in-array.php#example-6324), etc (probably everything that can be an array element), while `isset` can only handle elements that are valid as array keys (basically `int`s and `string`s)...