@Ekman 2020-06-17
@jdreesen That's a fair point.

I propose one of these solutions:

1. Use `isset` and introduce a backwards incompatible change
2. Use `SplObjectStorage` for checking distinct objects. Use an `array` to check distinct primitive values
3. Create a wrapper for primitive values so that everything can be stored in a `SplObjectStorage`
4. Keep using `in_array`

It would be a shame to keep the solution as-is. Decreasing the time complexity from `O(n^2)` to `O(n)` is a huge improvement, even if it requires some weird code. This is truly the example where the benefits of optimized code outweighs readable code. :)